[FUNCTOR-3] Reduced the use of raw types in test classes by adding generics where needed. In some cases, specially when using arrays with generics, the warnings were suppressed.
This commit also fixed some warnings in Eclipse. In TestAbstractLoopProcedure.MockLoopProcedure class, only the serialVersionUID was included, removing a warning. And in TestLazyMap#testBaseMapOverrides, instead of using a pair <String, String>, the test was updated to use a <String, Integer>. Although the data used in the tests is different, the behavior is still the same. And in TestIsWithinRange, one unused import was removed.

All tests passed locally, no CPD, PMD or CheckStyle problems found.

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/functor/trunk@1363514 13f79535-47bb-0310-9956-ffa450edef68