[FUNCTOR-12] Added more tests for the adapter classes. Also removed unreachable code, as it could not be tested and was immutable and created/validated during construction.
Some objects were immutable but could be null. In cases like this, no code was added, and the null validation was left in place, i.e. no if's were removed. See FullyBoundFunction for an example where the functor was validated, but the other arguments could be null, and in equals() and hashcode() the validation wasn't removed.

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/functor/trunk@1365377 13f79535-47bb-0310-9956-ffa450edef68