Removed the "stoppable" behavior from the Generator interface, and also removed the wrapped generator from BaseGenerator. Both were moved to a new abstract class, LoopGenerator.
The Generator of [functor] is responsible for executing a procedure for each 
element that is generates. However, there are some implementations of the 
Generator interface in [functor] that require special behavior (stoppable and 
wrapped generator). 

Moving this behaviors to a new class, now one can implement a new generator 
without having to implement stop (Liskov principle).

The implementations that control the execution flow were moved to the loop 
package. The existing IntegerRange and LongRange were moved to a range package, 
and an utility class Ranges has been created, but hasn't been implemented yet. 
This class will create ranges (that are generators).

TransformedGenerator has remarks about the suppression of warnings. This is a 
special generator, that wraps a generator, but returns another one. So it 
breaks the interface contract, and we suppress the warnings when we cast the 
wrapped generator. This class has been marked as final, to avoid bogus 
specializations.

The simplification of the Generator interface may be specially useful when 
separating classes for functor-api and functor-impl.

Tests were updated accordingly.

git-svn-id: https://svn.apache.org/repos/asf/commons/proper/functor/branches/generators-FUNCTOR-14@1374881 13f79535-47bb-0310-9956-ffa450edef68